+++ spring-boot-email-core/src/test/java/it/ozimov/springboot/templating/mail/service/ApplicationPropertiesConstantsTest.java
@@ -21,15 +21,16 @@ public class ApplicationPropertiesConstantsTest {
         assertions.assertThat(SPRING_MAIL_PROPERTIES_MAIL_SMTP_AUTH).isEqualTo("spring.mail.properties.mail.smtp.auth");
         assertions.assertThat(SPRING_MAIL_PROPERTIES_MAIL_SMTP_STARTTLS_ENABLE).isEqualTo("spring.mail.properties.mail.smtp.starttls.enable");
         assertions.assertThat(SPRING_MAIL_PROPERTIES_MAIL_SMTP_STARTTLS_REQUIRED).isEqualTo("spring.mail.properties.mail.smtp.starttls.required");
+        assertions.assertThat(SPRING_MAIL_SCHEDULER_ENABLED).isEqualTo("spring.mail.scheduler.enabled");
         assertions.assertThat(SPRING_MAIL_SCHEDULER_PRIORITY_LEVELS).isEqualTo("spring.mail.scheduler.priorityLevels");
-        assertions.assertThat(SPRING_MAIL_SCHEDULER_PERSISTENCE_LAYER_DESIRED_BATCH_SIZE).isEqualTo("spring.mail.scheduler.persistenceLayer.desiredBatchSize");
-        assertions.assertThat(SPRING_MAIL_SCHEDULER_PERSISTENCE_LAYER_MIN_KEPT_IN_MEMORY).isEqualTo("spring.mail.scheduler.persistenceLayer.minKeptInMemory");
-        assertions.assertThat(SPRING_MAIL_SCHEDULER_PERSISTENCE_LAYER_MAX_KEPT_IN_MEMORY).isEqualTo("spring.mail.scheduler.persistenceLayer.maxKeptInMemory");
-        assertions.assertThat(SPRING_MAIL_PERSISTENCE_ENABLED).isEqualTo("spring.mail.persistence.enabled");
-        assertions.assertThat(SPRING_MAIL_PERSISTENCE_REDIS_ENABLED).isEqualTo("spring.mail.persistence.redis.enabled");
-        assertions.assertThat(SPRING_MAIL_PERSISTENCE_REDIS_EMBEDDED).isEqualTo("spring.mail.persistence.redis.embedded");
-        assertions.assertThat(SPRING_MAIL_PERSISTENCE_REDIS_HOST).isEqualTo("spring.mail.persistence.redis.host");
-        assertions.assertThat(SPRING_MAIL_PERSISTENCE_REDIS_PORT).isEqualTo("spring.mail.persistence.redis.port");
+        assertions.assertThat(SPRING_MAIL_PERSISTENCE_ENABLED).isEqualTo("spring.mail.scheduler.persistence.enabled");
+        assertions.assertThat(SPRING_MAIL_SCHEDULER_PERSISTENCE_DESIRED_BATCH_SIZE).isEqualTo("spring.mail.scheduler.persistence.desiredBatchSize");
+        assertions.assertThat(SPRING_MAIL_SCHEDULER_PERSISTENCE_MIN_KEPT_IN_MEMORY).isEqualTo("spring.mail.scheduler.persistence.minKeptInMemory");
+        assertions.assertThat(SPRING_MAIL_SCHEDULER_PERSISTENCE_MAX_KEPT_IN_MEMORY).isEqualTo("spring.mail.scheduler.persistence.maxKeptInMemory");
+        assertions.assertThat(SPRING_MAIL_PERSISTENCE_REDIS_ENABLED).isEqualTo("spring.mail.scheduler.persistence.redis.enabled");
+        assertions.assertThat(SPRING_MAIL_PERSISTENCE_REDIS_EMBEDDED).isEqualTo("spring.mail.scheduler.persistence.redis.embedded");
+        assertions.assertThat(SPRING_MAIL_PERSISTENCE_REDIS_HOST).isEqualTo("spring.mail.scheduler.persistence.redis.host");
+        assertions.assertThat(SPRING_MAIL_PERSISTENCE_REDIS_PORT).isEqualTo("spring.mail.scheduler.persistence.redis.port");
     }
 
 }