+++ spring-boot-email-template-engines/spring-boot-thymeleaf-email/src/main/java/it/ozimov/springboot/templating/mail/service/ThymeleafTemplateService.java
@@ -22,7 +22,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.thymeleaf.context.Context;
-import org.thymeleaf.context.WebContext;
 import org.thymeleaf.spring4.SpringTemplateEngine;
 
 import java.io.IOException;